+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Preference system
 - Touch support
+- Fullscreen toggle in editor
 
 ## [1.0.0-alpha.1] - 2024-10-27